Move FFXIV results to components folder and give dark theme
