+++ app/components/FFXIVResults/FullScan/DraggableHeader.tsx
