+++ app/components/FFXIVResults/FullScan/getOrderedColumns.ts
