+++ app/components/form/ffxiv/CheckBoxModal.tsx
