+++ app/routes/queries/previous-search/index.tsx
@@ -3,7 +3,7 @@ import { json } from '@remix-run/cloudflare'
 import { useLoaderData } from '@remix-run/react'
 import { useTypedSelector } from '~/redux/useTypedSelector'
 import NoResults from './NoResults'
-import Results from '../full-scan/Results'
+import Results from '../../../components/FFXIVResults/FullScan/Results'
 
 export const loader: LoaderFunction = ({ request }) => {
   const url = new URL(request.url)
