+++ app/tailwind.css
@@ -779,6 +779,14 @@ select {
   top: 0px;
 }
 
+.-right-1 {
+  right: -0.25rem;
+}
+
+.left-0 {
+  left: 0px;
+}
+
 .top-5 {
   top: 1.25rem;
 }
@@ -807,14 +815,6 @@ select {
   right: 0px;
 }
 
-.-right-1 {
-  right: -0.25rem;
-}
-
-.left-0 {
-  left: 0px;
-}
-
 .z-50 {
   z-index: 50;
 }
@@ -827,16 +827,16 @@ select {
   z-index: 10000;
 }
 
+.z-\[inherit\] {
+  z-index: inherit;
+}
+
 .z-40 {
   z-index: 40;
 }
 
 .z-10 {
   z-index: 10;
-}
-
-.z-\[inherit\] {
-  z-index: inherit;
 }
 
 .col-span-6 {
@@ -980,12 +980,12 @@ select {
   margin-bottom: 0.25rem;
 }
 
-.mt-0\.5 {
-  margin-top: 0.125rem;
-}
-
 .mt-0 {
   margin-top: 0px;
+}
+
+.mt-0\.5 {
+  margin-top: 0.125rem;
 }
 
 .mr-3 {
@@ -1174,6 +1174,10 @@ select {
   width: 0px;
 }
 
+.w-\[95\%\] {
+  width: 95%;
+}
+
 .w-10 {
   width: 2.5rem;
 }
@@ -1196,10 +1200,6 @@ select {
   width: 16rem;
 }
 
-.w-\[95\%\] {
-  width: 95%;
-}
-
 .w-4\/5 {
   width: 80%;
 }
@@ -1218,12 +1218,12 @@ select {
   min-width: fit-content;
 }
 
-.min-w-\[150px\] {
-  min-width: 150px;
-}
-
 .min-w-\[100px\] {
   min-width: 100px;
+}
+
+.min-w-\[150px\] {
+  min-width: 150px;
 }
 
 .min-w-max {
@@ -1290,12 +1290,12 @@ select {
   flex-grow: 1;
 }
 
-.grow-0 {
-  flex-grow: 0;
-}
-
 .grow {
   flex-grow: 1;
+}
+
+.grow-0 {
+  flex-grow: 0;
 }
 
 .basis-52 {
@@ -1783,6 +1783,11 @@ select {
   background-color: rgb(55 65 81 / var(--tw-bg-opacity));
 }
 
+.bg-blue-800 {
+  --tw-bg-opacity: 1;
+  background-color: rgb(30 64 175 / var(--tw-bg-opacity));
+}
+
 .bg-gray-500 {
   --tw-bg-opacity: 1;
   background-color: rgb(107 114 128 / var(--tw-bg-opacity));
@@ -1801,11 +1806,6 @@ select {
 .bg-gray-900 {
   --tw-bg-opacity: 1;
   background-color: rgb(17 24 39 / var(--tw-bg-opacity));
-}
-
-.bg-blue-800 {
-  --tw-bg-opacity: 1;
-  background-color: rgb(30 64 175 / var(--tw-bg-opacity));
 }
 
 .bg-blue-100 {
@@ -1910,14 +1910,14 @@ select {
   padding-bottom: 1.25rem;
 }
 
-.py-0 {
-  padding-top: 0px;
-  padding-bottom: 0px;
-}
-
 .py-4 {
   padding-top: 1rem;
   padding-bottom: 1rem;
+}
+
+.py-0 {
+  padding-top: 0px;
+  padding-bottom: 0px;
 }
 
 .pb-2 {
@@ -2147,6 +2147,10 @@ select {
   opacity: 0;
 }
 
+.opacity-80 {
+  opacity: 0.8;
+}
+
 .opacity-75 {
   opacity: 0.75;
 }
@@ -2157,10 +2161,6 @@ select {
 
 .opacity-100 {
   opacity: 1;
-}
-
-.opacity-80 {
-  opacity: 0.8;
 }
 
 .shadow {
@@ -2460,6 +2460,21 @@ select {
   display: flex;
 }
 
+.dark .dark\:divide-gray-400 > :not([hidden]) ~ :not([hidden]) {
+  --tw-divide-opacity: 1;
+  border-color: rgb(156 163 175 / var(--tw-divide-opacity));
+}
+
+.dark .dark\:divide-gray-700 > :not([hidden]) ~ :not([hidden]) {
+  --tw-divide-opacity: 1;
+  border-color: rgb(55 65 81 / var(--tw-divide-opacity));
+}
+
+.dark .dark\:divide-gray-600 > :not([hidden]) ~ :not([hidden]) {
+  --tw-divide-opacity: 1;
+  border-color: rgb(75 85 99 / var(--tw-divide-opacity));
+}
+
 .dark .dark\:border-0 {
   border-width: 0px;
 }
@@ -2485,11 +2500,6 @@ select {
 .dark .dark\:border-gray-200 {
   --tw-border-opacity: 1;
   border-color: rgb(229 231 235 / var(--tw-border-opacity));
-}
-
-.dark .dark\:border-gray-900 {
-  --tw-border-opacity: 1;
-  border-color: rgb(17 24 39 / var(--tw-border-opacity));
 }
 
 .dark .dark\:border-gray-500 {
@@ -2532,18 +2542,22 @@ select {
   background-color: rgb(59 130 246 / var(--tw-bg-opacity));
 }
 
-.dark .dark\:bg-transparent {
-  background-color: transparent;
+.dark .dark\:bg-gray-800 {
+  --tw-bg-opacity: 1;
+  background-color: rgb(31 41 55 / var(--tw-bg-opacity));
 }
 
-.dark .dark\:bg-gray-900 {
-  --tw-bg-opacity: 1;
-  background-color: rgb(17 24 39 / var(--tw-bg-opacity));
+.dark .dark\:bg-transparent {
+  background-color: transparent;
 }
 
 .dark .dark\:bg-gray-500 {
   --tw-bg-opacity: 1;
   background-color: rgb(107 114 128 / var(--tw-bg-opacity));
+}
+
+.dark .dark\:font-medium {
+  font-weight: 500;
 }
 
 .dark .dark\:text-gray-100 {
@@ -2604,24 +2618,14 @@ select {
   --quick-links-hover-bg: #1e293b;
 }
 
-.dark .dark\:hover\:bg-gray-700:hover {
+.dark .dark\:hover\:bg-gray-400:hover {
   --tw-bg-opacity: 1;
-  background-color: rgb(55 65 81 / var(--tw-bg-opacity));
+  background-color: rgb(156 163 175 / var(--tw-bg-opacity));
 }
 
 .dark .dark\:hover\:bg-slate-800:hover {
   --tw-bg-opacity: 1;
   background-color: rgb(30 41 59 / var(--tw-bg-opacity));
-}
-
-.dark .dark\:hover\:bg-gray-500:hover {
-  --tw-bg-opacity: 1;
-  background-color: rgb(107 114 128 / var(--tw-bg-opacity));
-}
-
-.dark .dark\:hover\:bg-gray-400:hover {
-  --tw-bg-opacity: 1;
-  background-color: rgb(156 163 175 / var(--tw-bg-opacity));
 }
 
 .dark .dark\:hover\:text-blue-100:hover {
@@ -2642,6 +2646,16 @@ select {
 .dark .group:hover .dark\:group-hover\:border-gray-100 {
   --tw-border-opacity: 1;
   border-color: rgb(243 244 246 / var(--tw-border-opacity));
+}
+
+.dark .group:hover .dark\:group-hover\:bg-gray-600 {
+  --tw-bg-opacity: 1;
+  background-color: rgb(75 85 99 / var(--tw-bg-opacity));
+}
+
+.dark .group:hover .dark\:group-hover\:text-gray-200 {
+  --tw-text-opacity: 1;
+  color: rgb(229 231 235 / var(--tw-text-opacity));
 }
 
 .dark .group:hover .dark\:group-hover\:text-gray-100 {
